feature: add search exh card style
