--- .env 2
@@ -1,1 +1,0 @@
-VITE_firebaseApiKey=AIzaSyBqBIQPkKY2-4UL0jzwp1_aaH912cSV0VU
