+++ src/views/SearchExhibition.vue
@@ -1,238 +1,13 @@
 <template>
     <div class="d-flex">
         <!-- filter side -->
-        <div :class="`side border-end ${isFilterShow ? 'w-0' : 'border-none'}`">
-            <div :class="`m-4 ${isFilterShow && 'd-none'}`">
-                <div class="w-100 mb-2 d-block d-md-none">
-                    <button
-                        class="btn border"
-                        @click="filterShowHandler"
-                    >
-                        <CloseIcon />
-                        <span> Close</span>
-                    </button>
-                </div>
-                <!-- header -->
-                <div
-                    @click="filterSetHandler('dateValid', !filterSetData.dateValid)"
-                    class="d-flex gap-2 align-items-center border-bottom pb-4 mb-4 px-3"
-                >
-                    <input
-                        id="date-valid"
-                        type="checkbox"
-                        name="checkbox"
-                        :checked="filterSetData.dateValid"
-                    />
-                    <!-- :checked="" -->
-                    <p
-                        for="date-valid"
-                        class="pb-1"
-                    >
-                        顯示已結束展覽
-                    </p>
-                </div>
-                <!-- Type -->
-                <div class="mb-3">
-                    <p class="fw-bold mb-2">類型</p>
-                    <!-- component radioList -->
-                    <ul class="type-list rounded-3 border mb-1">
-                        <li class="border-bottom p-2">
-                            <div @click="filterSetHandler('type', 'all')">
-                                <input
-                                    class="me-2"
-                                    id="type-all"
-                                    type="radio"
-                                    name="type"
-                                    value="all"
-                                    :checked="filterSetData.type === 'all'"
-                                />
-                                <span for="type-all">全部</span>
-                            </div>
-                        </li>
-                        <li
-                            class="p-2 border-bottom"
-                            v-for="(item, index) in filterData.typeArr.slice(
-                                0,
-                                filterData.typeQuantity
-                            )"
-                            :key="index"
-                        >
-                            <div @click="filterSetHandler('type', item)">
-                                <input
-                                    class="me-2"
-                                    :id="item"
-                                    type="radio"
-                                    name="type"
-                                    :value="item"
-                                    :checked="filterSetData.type === item"
-                                />
-                                <span
-                                    class=""
-                                    :for="item"
-                                >
-                                    {{ item }}</span
-                                >
-                            </div>
-                        </li>
-                    </ul>
-                    <div class="d-flex px-1">
-                        <button
-                            :class="`border-0 bg-transparent fw-bold ${
-                                filterData.typeQuantity < 6 && 'd-none'
-                            }`"
-                            type="button"
-                            @click="typeQtyHandler('reset')"
-                        >
-                            收回
-                        </button>
-                        <button
-                            @click="typeQtyHandler('add')"
-                            :class="`border-0 bg-transparent fw-bold ms-auto ${
-                                filterData.typeArr.length < 5 && 'd-none'
-                            }`"
-                            type="button"
-                        >
-                            顯示更多
-                        </button>
-                    </div>
-                </div>
-                <!-- Time -->
-                <div class="mb-2">
-                    <p class="fw-bold mb-2">展覽期間</p>
-                    <DatePicker
-                        v-model="filterSetData.startDate"
-                        class="mb-2"
-                        type="number"
-                        placeholder="開始時間"
-                    />
-
-                    <!--  v-model="filterSetData.startDate" -->
-                    <DatePicker
-                        placeholder="結束時間"
-                        v-model="filterSetData.endDate"
-                    />
-                </div>
-                <!-- price two range -->
-                <p class="mb-2 fw-bold">價格</p>
-                <VeeForm
-                    v-slot="{ errors }"
-                    @submit="submit"
-                >
-                    <div class="d-flex flex-column">
-                        <p class="mb-1 fs-7">最小票價</p>
-                        <VeeField
-                            v-model="filterSetData.minPrice"
-                            :class="`border rounded py-1 px-2 ${
-                                errors[`最小價格`] ? 'border-danger' : 'border'
-                            }`"
-                            placeholder="最小價格"
-                            name="最小價格"
-                            type="number"
-                            oninput="value=value.replace(/[^\d]/g,'')"
-                            :rules="{
-                                min_value: 0,
-                                max_value: filterSetData.maxPrice,
-                                numeric: true
-                            }"
-                        />
-                        <ErrorMessage
-                            v-if="errors[`最小價格`]"
-                            as="p"
-                            class="text-danger errorMessage"
-                            name="最小價格"
-                        />
-                        <p
-                            v-else
-                            class="errorMessage"
-                        ></p>
-                    </div>
-                    <div class="d-flex flex-column">
-                        <p class="mb-1 fs-7">最大票價</p>
-                        <VeeField
-                            v-model="filterSetData.maxPrice"
-                            :class="`border rounded py-1 px-2 ${
-                                errors[`最大價格`] ? 'border-danger' : 'border'
-                            }`"
-                            placeholder="最大價格"
-                            name="最大價格"
-                            type="number"
-                            oninput="value=value.replace(/[^\d]/g,'')"
-                            :rules="{
-                                min_value: filterSetData.minPrice,
-                                max_value: 100000,
-                                numeric: true
-                            }"
-                        />
-                        <ErrorMessage
-                            v-if="errors[`最大價格`]"
-                            as="p"
-                            class="text-danger errorMessage"
-                            name="最大價格"
-                        />
-                        <p
-                            v-else
-                            class="errorMessage"
-                        ></p>
-                    </div>
-                </VeeForm>
-                <!-- 地區 -->
-                <div>
-                    <p class="fs-7 mb-2 fw-bold">城市 {{ filterData.cityKeyWord }}</p>
-                    <input
-                        @input="searchCityHandler"
-                        class="border w-100 rounded px-2 py-1 mb-2"
-                        type="search"
-                        placeholder="篩選城市名稱"
-                        name="cityKeyWord"
-                        id="cityKeyWord"
-                    />
-                    <ul class="city-list border rounded-3">
-                        <li class="p-2 border-bottom">
-                            <div @click="filterSetHandler('city', 'all')">
-                                <input
-                                    class="me-2"
-                                    id="all"
-                                    type="radio"
-                                    name="city"
-                                    value="all"
-                                    :checked="filterSetData.city === 'all'"
-                                />
-                                <span for="all">全部</span>
-                            </div>
-                        </li>
-                        <li
-                            class="p-2 border-bottom"
-                            v-for="(item, index) in filterData.cityArr"
-                            :key="index"
-                        >
-                            <div @click="filterSetHandler('city', item)">
-                                <input
-                                    class="me-2"
-                                    :id="item"
-                                    type="radio"
-                                    name="city"
-                                    :value="item"
-                                    :checked="filterSetData.city === item"
-                                />
-                                <span
-                                    class=""
-                                    :for="item"
-                                >
-                                    {{ item }}
-                                </span>
-                            </div>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </div>
+        <Filter
+            :is-filter-show="isFilterShow"
+            :filter-show-handler="filterShowHandler"
+        />
         <!-- main -->
         <div class="container position-relative px-3">
-            <div class="border">
-                <p>開發資料</p>
-                <p>req query 檢查資料 {{ route.query }}</p>
-            </div>
-            <div class="my-4 rounded-3">
+            <div class="mt-5 rounded-3">
                 <!-- 展覽搜索 -->
                 <div>
                     <input
@@ -243,174 +18,143 @@
                         id=""
                     />
                 </div>
+                {{ queryData }}
             </div>
 
             <!-- :class="`btn  d-flex align-items-center gap-2  ${
                     isFilterShow ? 'btn-outline-dark' : 'btn-dark'
                 }`" -->
             <button
-                class="border position-absolute top-0 start-0"
+                class="btn position-absolute top-0 start-0"
                 type="button"
                 @click="filterShowHandler"
             >
                 <FilterMenu />
                 <span> 篩選器</span>
             </button>
-            <p>篩選器 {{ filterSetData }} 這裡需要設定 增加樣式</p>
-            <p></p>
-            <p>展覽資料 {{ store.exhibitionList.map((item) => item.name) }}</p>
+            <p>篩選器 這裡需要設定 增加樣式</p>
+            <div class="row">
+                <div class="col-12 col-md-6 col-lg-4">
+                    <div class="border position-relative shadow rounded-3 overflow-hidden">
+                        <!-- city -->
+                        <div
+                            class="position-absolute bg-secondary text-white px-3 py-2 border-bottom-end-3"
+                        >
+                            {{ store.exhibitionList.map((item) => item)[0]?.location.country }}
+                        </div>
+                        <!-- header -->
+                        <div class="rounded">
+                            <img
+                                class="object-fit-cover"
+                                :src="store.exhibitionList.map((item) => item)[0]?.image"
+                                alt=""
+                            />
+                        </div>
+                        <!-- exh content -->
+                        <div class="p-3">
+                            <!-- exh title  -->
+                            <div class="fw-bold fs-5">
+                                <p>{{ store.exhibitionList.map((item) => item)[0]?.name }}</p>
+                            </div>
+                            <!-- exh-date -->
+                            <div class="d-flex gap-2 mb-1">
+                                <span>{{
+                                    new Date(
+                                        store.exhibitionList.map((item) => item)[0]?.startDate
+                                    ).toLocaleDateString()
+                                }}</span>
+                                <span>~</span>
+                                <span>{{
+                                    new Date(
+                                        store.exhibitionList.map((item) => item)[0]?.endDate
+                                    ).toLocaleDateString()
+                                }}</span>
+                            </div>
+                            <!-- exh type -->
+                            <ul class="d-flex">
+                                <li
+                                    class="border py-1 px-2 bg-success text-white rounded-pill fs-7"
+                                >
+                                    {{ store.exhibitionList.map((item) => item)[0]?.type }}
+                                </li>
+                            </ul>
+                        </div>
+                    </div>
+                </div>
+            </div>
         </div>
     </div>
 </template>
 
 <script setup>
+    // import _ from 'lodash'
+    // import gsap from 'gsap'
+    import { ref, onMounted, watch } from 'vue'
+    import { useRoute } from 'vue-router'
+    import { exhibitionStore } from '../stores/exhibitionList'
+    import Filter from '../components/Filter.vue'
     // import TwoRangeSlider from '../components/TwoRangeSlider.vue'
 
-    import { ref, watch, onMounted } from 'vue'
-    import { exhibitionStore } from '../stores/exhibitionList'
-    import { useRoute, useRouter } from 'vue-router'
-    // import gsap from 'gsap'
+    const store = exhibitionStore()
+    const route = useRoute()
 
-    import _ from 'lodash'
+    const queryData = ref()
+    const filterExhData = ref()
 
-    const submit = (e) => {
-        e.preventDefault()
+    const queryDataView = (data) => {
+        //排除 初始數據做呈現
+        const { city, dateValid, startDate, endDate, maxPrice, minPrice, type } = data
+        const newValue = {
+            dateValid: dateValid === 'false' ? null : dateValid,
+            city: city === 'all' ? null : city,
+            startDate: startDate !== '0' ? parseInt(startDate) : null,
+            endDate: endDate !== '0' ? parseInt(endDate) : null,
+            maxPrice: maxPrice === '100000' ? null : parseInt(maxPrice),
+            minPrice: minPrice === '0' ? null : parseInt(minPrice),
+            type: type === 'all' ? null : type
+        }
+
+        Object.keys(newValue).forEach((keyName) => {
+            if (newValue[keyName] === null || newValue[keyName] === undefined)
+                delete newValue[keyName]
+
+            if (keyName === 'maxPrice' || keyName === 'minPrice') {
+                if (isNaN(newValue[keyName])) {
+                    delete newValue[keyName]
+                }
+            }
+        })
+
+        //price 沒有通過篩選 刪除
+        if (newValue.maxPrice < newValue.minPrice || newValue.maxPrice > 100000) {
+            delete newValue['maxPrice']
+            delete newValue['minPrice']
+        }
+
+        // console.log('queryDataView', newValue)
+        queryData.value = { ...newValue }
     }
 
+    onMounted(() => {
+        queryDataView(route.query)
+        //篩選data
+    })
+
+    watch(route, () => {
+        console.log('嘿嘿 , 網址更新囉', route.query)
+        queryDataView(route.query)
+        //篩選data
+    })
+
     //exhibition Data
-    const store = ref(exhibitionStore())
-
-    //url query
-    const route = useRoute()
-    const router = useRouter()
-
     ////////////filter////////////
-    //side bar show
     const isFilterShow = ref(false)
     const filterShowHandler = () => {
         isFilterShow.value = !isFilterShow.value
     }
-
-    const cityArr = [
-        '臺北',
-        '新北',
-        '桃園',
-        '臺中',
-        '臺南',
-        '高雄',
-        '新竹',
-        '苗栗',
-        '彰化',
-        '南投',
-        '雲林',
-        '嘉義',
-        '屏東',
-        '宜蘭',
-        '花蓮',
-        '臺東',
-        '澎湖',
-        '金門',
-        '連江',
-        '基隆'
-    ]
-
-    //filter data
-    const filterData = ref({
-        typeArr: [],
-        cityArr,
-        cityKeyWord: '',
-        typeQuantity: 5
-    })
-
-    //Filter setting Data
-    const filterSetData = ref({
-        dateValid: false,
-        type: 'all',
-        startDate: 0,
-        endDate: new Date().getTime() + 365 * 24 * 60 * 60 * 100,
-        minPrice: 1,
-        maxPrice: 100000,
-        city: 'all'
-    })
-
-    onMounted(async () => {
-        if (store.value.exhibitionList.length === 0) {
-            await store.value.getAllExhibitionData() // 取得展覽全部資料
-        }
-        initFilterData() //初始化篩選器資料
-        console.log('[BeforeMount]', store.value.exhibitionList)
-        console.log('[BeforeMount]', route.query)
-
-        watch(filterSetData.value, () => {
-            if (typeof filterSetData.value.startDate === 'object') {
-                filterSetData.value.startDate = new Date(filterSetData.value.startDate).getTime()
-            }
-            if (typeof filterSetData.value.endDate === 'object') {
-                filterSetData.value.endDate = new Date(filterSetData.value.endDate).getTime()
-            }
-            router.replace({
-                name: 'SearchExhibition',
-                query: filterSetData.value
-            })
-        })
-    })
-
-    //init Filter Data
-    const initFilterData = () => {
-        //init type data
-        filterData.value.typeArr = Array.from(
-            new Set(store.value.exhibitionList.map((item) => item.type))
-        )
-
-        //query setting
-        if (Object.keys(route.query).length > 0) {
-            const { dateValid, type, startDate, endDate, minPrice, maxPrice, city } = route.query
-            const reSetData = {
-                dateValid: dateValid === 'true',
-                type,
-                startDate: parseInt(startDate),
-                endDate: parseInt(endDate),
-                minPrice: parseInt(minPrice),
-                maxPrice: parseInt(maxPrice),
-                city
-            }
-            // //fix
-            filterSetData.value = { ...filterSetData.value, ...reSetData }
-        }
-    }
-
-    //[Filter setting Data] controller current type value
-    const filterSetHandler = (keyName, value) => {
-        console.log('click', keyName, value)
-
-        filterSetData.value[`${keyName}`] = value
-
-        console.log('[fetch Data]', 123)
-    }
-
-    //[Filter Data] controller
-    const searchCityHandler = (e) => {
-        e.preventDefault()
-        filterData.value.cityKeyWord = e.target.value
-        if (e.target.value === '') filterData.value.cityArr = cityArr
-        console.log(filterData.value.cityArr.filter((item) => item.includes(e.target.value)))
-        filterData.value.cityArr = cityArr.filter((item) => item.includes(e.target.value))
-    }
-
-    //[Filter type] show Quantity
-    const typeQtyHandler = (controller) => {
-        switch (controller) {
-            case 'add':
-                filterData.value.typeQuantity += 5
-                break
-            case 'reset':
-                filterData.value.typeQuantity = 5
-                break
-        }
-    }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
     //filter slider
     .side {
         overflow: hidden;
@@ -447,5 +191,13 @@
 
     .w-0 {
         min-width: 0;
+    }
+
+    .border-top-start-1 {
+        border-top-left-radius: 0.5rem;
+    }
+
+    .border-bottom-end-3 {
+        border-bottom-right-radius: 1.5rem;
     }
 </style>
