+++ vite.config.js
@@ -17,7 +17,7 @@ export default defineConfig({
     },
     server: {
         // 啟動 server 時預設開啟的頁面
-        open: '/'
+        // open: '/'
     },
     build: {
         outDir: 'dist'
